수정: 지수표현된 수를 long 타입으로 변환 못 하는 문제, BigInteger 사용 못 하는 문제 해결

+++ src/main/java/com/clipsoft/cson/util/NumberConversionUtil.java
@@ -166,7 +166,7 @@ public class NumberConversionUtil {
             if(bi.bitLength() <= 63){
                 return bi.longValue();
             }
-            return bi;
+            return bi.toBigInteger();
         } else if(numberConversionOption.isIgnoreNonNumeric()) {
             return null;
         }
